+++ contracts/crowdsale/validation/WhitelistCrowdsale.sol
@@ -14,7 +14,7 @@ contract WhitelistCrowdsale is WhitelistedRole, Crowdsale {
      * @param _beneficiary Token beneficiary
      * @param _weiAmount Amount of wei contributed
      */
-    function _preValidatePurchase(address _beneficiary, uint256 _weiAmount) internal view {
+    function _preValidatePurchase(address _beneficiary, uint _weiAmount) internal view {
         require(isWhitelisted(_beneficiary), "WhitelistCrowdsale: beneficiary doesn't have the Whitelisted role");
         super._preValidatePurchase(_beneficiary, _weiAmount);
     }
