+++ contracts/cryptography/MerkleProof.sol
@@ -15,7 +15,7 @@ library MerkleProof {
     function verify(bytes32[] memory proof, bytes32 root, bytes32 leaf) internal pure returns (bool) {
         bytes32 computedHash = leaf;
 
-        for (uint256 i = 0; i < proof.length; i++) {
+        for (uint i = 0; i < proof.length; i++) {
             bytes32 proofElement = proof[i];
 
             if (computedHash <= proofElement) {
