+++ contracts/math/Math.sol
@@ -9,14 +9,14 @@ library Math {
     /**
      * @dev Returns the largest of two numbers.
      */
-    function max(uint256 a, uint256 b) internal pure returns (uint256) {
+    function max(uint a, uint b) internal pure returns (uint) {
         return a >= b ? a : b;
     }
 
     /**
      * @dev Returns the smallest of two numbers.
      */
-    function min(uint256 a, uint256 b) internal pure returns (uint256) {
+    function min(uint a, uint b) internal pure returns (uint) {
         return a < b ? a : b;
     }
 
@@ -24,7 +24,7 @@ library Math {
      * @dev Returns the average of two numbers. The result is rounded towards
      * zero.
      */
-    function average(uint256 a, uint256 b) internal pure returns (uint256) {
+    function average(uint a, uint b) internal pure returns (uint) {
         // (a + b) / 2 can overflow, so we distribute
         return (a / 2) + (b / 2) + ((a % 2 + b % 2) / 2);
     }
